Moved 404 & Profile stories to component for deploy
