+++ components/Profile/Profile.stories.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { RecoilRoot } from 'recoil'
 
-import Profile from './Profile'
+import Profile from '../../pages/profile/Profile'
 
 export default {
   title: 'Profile',
